Add AktionsfinderProspektFixer
For some reason the site, when showing the magazines, will default to
having an obnoxious 300px right padding. I assume for ads, but I use an
adblocker so no ads load in, but the canvas is still taken in by empty
space that I could use to better view the magazine.