Properly check for archive status before redirecting
I was under the assumption that Archive would return a 404 if the page
is not yet archived, but seems that's not the case. The site will always
return a 200, because in the case the page is not yet archived, it will
simply redirect you to the "archive it now" page, and that's a
successful response.

An easy way to tell if we're on the actual archived page, or the
"archive it now" page, is to look at the path of the URL. If the page is
archived, the `/newest/` will be replaced by a specific timestamp of
when that page was archived, like `/20250326055136/`. If it still has
`/newest/` in the path, it means it wasn't archived yet and we got
served the "archive it now" page.

Obviously another way of checking would be to parse the HTML, but this
way is faster as I can keep making only an OPTIONS request and not even
get the body at all.