DWLGN Remove IntersectionObserver
It was a cool concept, to dynamically detect the visibility (in the
screen, not just `visible` style) and float or not the audio control
bar.

However, the IntersectionObserver API relies on a parent node to check
the visibility against, and so it generates issues when you move the
target to some other position.

I could have fixed this by setting some other parent node, but in the
end the fixed bottom in all cases looks decent enough, and so I don't
need the complex code.